カメラの座標を整数&layer_and_noの名前を変更 Change camera coordinates to integers and

+++ src/Pictures/Camera.h
@@ -5,15 +5,15 @@
 namespace pictures{
     class Camera{
         private:
-        common::Vec2<double> m_xy; // カメラの座標  Camera Coordinate
+        common::Vec2<int32_t> m_xy; // カメラの座標  Camera Coordinate
         double m_zoom; // カメラのズーム率  zoom rate of camera
 
         public:
         Camera();
         // アクセサ  Accessor
-        common::Vec2<double> GetXY() const {return m_xy;}
+        common::Vec2<int32_t> GetXY() const {return m_xy;}
         double GetZoom() const {return m_zoom;}
-        void SetXY(common::Vec2<double> xy){m_xy = xy;}
+        void SetXY(common::Vec2<int32_t> xy){m_xy = xy;}
         void SetZoom(double zoom){m_zoom = zoom;}
     };
 }
